Add files via upload
It has one error that I couldn't find so far. The error is when user enter a birth day input which is a letter, program do not work properly(the program must force the user until s/he enter actual correct input) however, when the same process accomplish after user enter a input for birth year which is a letter it runs properly.